+++ src/fidesctl/core/utils.py
@@ -1,5 +1,6 @@
 """Utils to help with API calls."""
 import logging
+import re
 from functools import partial
 from json.decoder import JSONDecodeError
 from typing import Dict, Iterator
@@ -10,7 +11,8 @@ import requests
 import sqlalchemy
 from sqlalchemy.engine import Engine
 
-from fideslang.models import DatasetField
+from fideslang.models import DatasetField, FidesModel
+from fideslang.validation import FidesValidationError
 
 logger = logging.getLogger("server_api")
 
@@ -80,3 +82,25 @@ def get_all_level_fields(fields: list) -> Iterator[DatasetField]:
         if field.fields:
             for nested_field in get_all_level_fields(field.fields):
                 yield nested_field
+
+
+def check_fides_key(proposed_fides_key: str) -> str:
+    """
+    A helper function to automatically sanitize
+    an invalid FidesKey.
+    """
+    try:
+        FidesModel(fides_key=proposed_fides_key)
+        return proposed_fides_key
+    except FidesValidationError as error:
+        echo_red(error)
+        return sanitize_fides_key(proposed_fides_key)
+
+
+def sanitize_fides_key(proposed_fides_key: str) -> str:
+    """
+    Attempts to manually sanitize a fides key if restricted
+    characters are discovered.
+    """
+    sanitized_fides_key = re.sub(r"[^a-zA-Z0-9_.-]", "_", proposed_fides_key)
+    return sanitized_fides_key