+++ CHANGELOG.md
@@ -25,6 +25,7 @@ The types of changes are:
 ### Fixed
 
 * Resolved a failure with populating applicable data subject rights to a data map
+* Handle invalid characters when generating a `fides_key` [#761](https://github.com/ethyca/fides/pull/761)
 
 ## [1.6.0](https://github.com/ethyca/fides/compare/1.5.3...1.6.0) - 2022-05-02